// src/nodes/schemas/t7/gen/models/NewAuthor.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type NewAuthor = {
    readonly type: string;
    readonly id: string;
    readonly host: string;
    displayName?: string;
    readonly url: string;
    github?: string;
    profileImage?: string;
    username: string;
    password: string;
};